+++ lib/dimigoin_flutter_plugin.dart
@@ -14,6 +14,7 @@ part 'models/dimigoin_user.dart';
 part 'models/dalgeurak_warning.dart';
 part 'models/dalgeurak_mealcancel.dart';
 part 'models/dalgeurak_mealexception.dart';
+part 'models/dalgeurak_conveniencefood.dart';
 part 'models/stream_socket.dart';